feat: errors handling

+++ server/src/routes/clientsRouter.ts
@@ -1,13 +1,15 @@
 import { clientSchema, clientUpdateSchema } from '../models/client'
+import { Client, Reservation } from '@prisma/client'
 import { Router, Request, Response } from 'express'
 import prisma from '../lib/prismaClient'
-import { Client } from '@prisma/client'
 import { sleep } from '../lib/sleep'
+import { ApiError } from '../lib/ApiError'
 
 interface ResponseEntity {
    message: string
    client?: Client
    clients?: Client[]
+   reservations?: Reservation[]
    error?: unknown
 }
 
@@ -83,7 +85,7 @@ clientsRouter.put('/:id', async (req: Request, res: Response<ResponseEntity>) =>
    }
 })
 
-clientsRouter.delete('/:id', async (req: Request, res: Response<ResponseEntity>) => {
+clientsRouter.delete('/:id', async (req: Request, res: Response) => {
    try {
       const { id } = req.params
 
@@ -92,21 +94,36 @@ clientsRouter.delete('/:id', async (req: Request, res: Response<ResponseEntity>)
       })
 
       if (!clientToDelete) {
-         res.status(404).send({ message: 'Cliente no encontrado' })
-         return
+         throw new ApiError(404, 'CLIENT_NOT_FOUND')
+      }
+
+      const clientReservations = await prisma.reservation.findMany({
+         where: { ownerId: id },
+      })
+
+      if (clientReservations.length > 0) {
+         throw new ApiError(400, 'CLIENT_HAS_RESERVATIONS', { reservations: clientReservations })
       }
 
       await prisma.client.delete({ where: { id } })
 
       await sleep(2000)
 
-      res.status(200).send({ message: 'Cliente eliminado', client: clientToDelete })
+      return res.status(200).send({ message: 'Cliente eliminado', client: clientToDelete })
    } catch (error: any) {
       console.log(error)
-      res.status(500).send({
-         message: 'Error eliminando el cliente',
-         error,
-      })
+      if (error instanceof ApiError) {
+         return res.status(error.statusCode).send({
+            errorCode: error.errorCode,
+            ...error.data,
+         })
+      }
+
+      return res.status(500).send({ code: 'INTERNAL_SERVER_ERROR' })
+      // return res.status(500).send({
+      //    message: 'Error eliminando el cliente',
+      //    error,
+      // })
    }
 })
